refactor: move config files to new folder
