+++ src/apps/ScopeChangeRequest/ScopeChangeRequestApp.tsx
@@ -2,12 +2,10 @@ import { ClientApi } from '@equinor/lighthouse-portal-client';
 import { setupWorkspaceSidesheet } from '../../Core/WorkSpace/src/WorkSpaceApi/Functions/setupWorkspaceSidesheet';
 import { SidesheetWrapper } from './Components/Sidesheet/SidesheetWrapper/SidesheetWrapper';
 import { ScopeChangeRequest } from './types/scopeChangeRequest';
-import { dataSource, idResolver } from './workspaceConfig/dataOptions';
 import { filterConfig } from './workspaceConfig/filter/filterConfig';
-import { prefetchQueriesOptions } from './workspaceConfig/prefetchQueryOptions';
 import { gardenConfig } from './workspaceConfig/sGarden/gardenConfig';
 import { tableConfig } from './workspaceConfig/sTable/tableConfig';
-import { statusBarConfig } from './utils/config';
+import { statusBarConfig, dataSource, idResolver, prefetchQueriesOptions } from './utils/config';
 
 const sidesheetCreator = setupWorkspaceSidesheet<ScopeChangeRequest, 'change'>({
     id: 'change',
