+++ src/apps/ScopeChangeRequest/utils/config/index.ts
@@ -1,1 +1,4 @@
 export * from './statusBarConfig';
+export * from './dataCreatorConfig';
+export * from './dataOptions';
+export * from './prefetchQueryOptions';
