+++ src/apps/ScopeChangeRequest/utils/config/prefetchQueryOptions.ts
@@ -1,0 +1,3 @@
+import { PrefetchQueriesOptions } from '../../../../Core/WorkSpace/src/WorkSpaceApi/workspaceState';
+
+export const prefetchQueriesOptions: PrefetchQueriesOptions[] = [];
