--- src/apps/ScopeChangeRequest/workspaceConfig/prefetchQueryOptions.ts
@@ -1,3 +1,0 @@
-import { PrefetchQueriesOptions } from '../../../Core/WorkSpace/src/WorkSpaceApi/workspaceState';
-
-export const prefetchQueriesOptions: PrefetchQueriesOptions[] = [];
